reorganize indexer code
