+++ src/types.ts
@@ -1,18 +1,27 @@
-import { Abi, ExtractAbiEventNames } from "abitype";
+import {
+  Abi,
+  AbiFunction,
+  AbiParametersToPrimitiveTypes,
+  ExtractAbiEventNames,
+  ExtractAbiFunction,
+  ExtractAbiFunctionNames,
+} from "abitype";
 import { GetEventArgs } from "viem";
-import type { Indexer } from "@/index";
+import type { CreateSubscriptionOptions } from "@/indexer";
 
 export type Hex = `0x${string}`;
 
 export type EventHandler<
-  TAbi extends Abi,
+  TAbi extends Abi = Abi,
   N extends ExtractAbiEventNames<TAbi> = ExtractAbiEventNames<TAbi>,
   TContext = unknown,
   TAbis extends Record<string, Abi> = Record<string, Abi>
 > = (args: {
   context: TContext;
-  readContract: Indexer<TAbis, TContext>["readContract"];
-  subscribeToContract: Indexer<TAbis, TContext>["subscribeToContract"];
+  chainId: number;
+  subscribeToContract: (
+    opts: Omit<CreateSubscriptionOptions<keyof TAbis>, "fromBlock">
+  ) => void;
   event: BaseEvent<
     N,
     GetEventArgs<
@@ -21,6 +30,24 @@ export type EventHandler<
       { EnableUnion: false; IndexedOnly: false; Required: true }
     >
   >;
+
+  readContract<
+    TContractName extends keyof TAbis,
+    TAbi extends Abi = TAbis[TContractName],
+    TFunctionName extends ExtractAbiFunctionNames<
+      TAbi,
+      "pure" | "view"
+    > = ExtractAbiFunctionNames<TAbi, "pure" | "view">,
+    TAbiFunction extends AbiFunction = ExtractAbiFunction<TAbi, TFunctionName>,
+    TReturn = AbiParametersToPrimitiveTypes<TAbiFunction["outputs"], "outputs">
+  >(args: {
+    contract: TContractName | keyof TAbis;
+    address: Hex;
+    functionName:
+      | TFunctionName
+      | ExtractAbiFunctionNames<TAbi, "pure" | "view">;
+    args?: AbiParametersToPrimitiveTypes<TAbiFunction["inputs"], "inputs">;
+  }): Promise<TReturn extends readonly [infer inner] ? inner : TReturn>;
 }) => Promise<void>;
 
 export type EventHandlers<
@@ -51,3 +78,18 @@ type BaseEvent<N = string, P = Record<string, unknown>> = {
 };
 
 export type ToBlock = "latest" | bigint;
+
+export type Contract<
+  TAbi extends Abi = Abi,
+  TContext = unknown,
+  TAbis extends Record<string, Abi> = Record<string, Abi>,
+  N extends ExtractAbiEventNames<TAbi> = ExtractAbiEventNames<TAbi>
+> = {
+  abi: TAbi;
+  subscriptions?: {
+    address: Hex;
+    fromBlock?: bigint;
+    toBlock?: ToBlock;
+  }[];
+  handlers?: Partial<EventHandlers<TAbi, N, TContext, TAbis>>;
+};
